make getNewFilterState an util function
