+++ src/mobility/utils.ts
@@ -5,6 +5,7 @@ import {
   RentalUrisFragment,
 } from '@atb/api/types/generated/fragments/mobility-shared';
 import {
+  FormFactorFilterType,
   getVisibleRange,
   MapRegion,
   MobilityMapFilterType,
@@ -22,6 +23,7 @@ import {
 import {Language} from '@atb/translations';
 import {formatDecimalNumber} from '@atb/utils/numbers';
 import {enumFromString} from '@atb/utils/enum-from-string';
+import {MobilityOperatorType} from '@atb-as/config-specs/lib/mobility-operators';
 
 export const isScooter = (
   feature: Feature<Point> | undefined,
@@ -170,3 +172,42 @@ export const isShowAll = (
 
 export const toFormFactorEnum = (str: string): FormFactor =>
   enumFromString(FormFactor, str) || FormFactor.Other;
+
+export const getNewFilterState = (
+  isChecked: boolean,
+  selectedOperator: string,
+  currentFilter: FormFactorFilterType | undefined,
+  allOperators: MobilityOperatorType[],
+): FormFactorFilterType => {
+  if (isChecked) {
+    // Add checked operator to list
+    const operators = [...(currentFilter?.operators ?? []), selectedOperator];
+    // If all operators are checked, set 'showAll' to true, rather that having all operators explicitly in the list.
+    // This allows for showing operators that do not exist in the whitelist
+    return operators.length === allOperators.length
+      ? {
+          operators: [],
+          showAll: true,
+        }
+      : {
+          operators,
+          showAll: false,
+        };
+  }
+  // If only one operator exists, treat unselecting this as unselecting all
+  if (allOperators.length === 1) {
+    return {
+      operators: [],
+      showAll: false,
+    };
+  }
+  // If 'showAll' was true at the time of unchecking one, all other operators should be added to the list.
+  const operators = currentFilter?.showAll
+    ? allOperators.map((o) => o.id).filter((o) => o !== selectedOperator)
+    : currentFilter?.operators?.filter((o: string) => o !== selectedOperator) ??
+      [];
+  return {
+    operators,
+    showAll: false,
+  };
+};
